+++ bld/wgml/c/grefdict.c
@@ -40,7 +40,7 @@
 
 char    *get_refid_value( char *p, att_val_type *attr_val, char *refid )
 {
-    size_t  k;
+    unsigned    k;
 
     p = get_att_value( p, attr_val );
     if( attr_val->name == NULL ) {       // no valid id
