+++ bld/wgml/h/wgml.h
@@ -199,16 +199,16 @@ extern  mac_entry * find_macro( mac_dict dict, char const *name );
 
 
 /* gmemory.c                            */
-extern void         *mem_alloc( size_t size );
+extern void         *mem_alloc( unsigned size );
 extern char         *mem_strdup( const char *str );
-extern char         *mem_tokdup( const char *str, int size );
+extern char         *mem_tokdup( const char *str, unsigned size );
 extern void         mem_banner( void );
 extern void         mem_free( void * p );
 extern void         mem_init( void );
 extern void         mem_fini( void );
 extern unsigned long mem_get_peak_usage( void );
 extern void         mem_prt_curr_usage( void );
-extern void         *mem_realloc( void * p, size_t size );
+extern void         *mem_realloc( void * p, unsigned size );
 extern int          mem_validate( void );
 extern void         mem_prt_list( void );
